requirements.txt
numpy
pandas
matplotlib
seaborn
kagglehub
fastapi
uvicorn
pillow
pyyaml
python-multipart